+++ WinSMARTS/schedAlgo.cpp
@@ -1,17 +1,17 @@
 #include "WinSMARTS.h"
 
 
-int RoundRobin(WinSMARTS* SMARTS)
+size_t RoundRobin(WinSMARTS* SMARTS)
 {
 	if (SMARTS->getTotalTasks() == 1)
 		return 0;
 
-	int nextTask = SMARTS->getCurrentTask() + 1;
+	size_t nextTask = SMARTS->getCurrentTask() + 1;
 	if(nextTask == SMARTS->getTotalTasks())
 		nextTask = 1;
 
 	//passes on all the tasks array and checks if there is ready task 
-	for(int count=1; SMARTS->getTaskStatus(nextTask) != READY && count<SMARTS->getTotalTasks(); ++count)
+	for(size_t count=1; SMARTS->getTaskStatus(nextTask) != READY && count < SMARTS->getTotalTasks(); ++count)
 	{
 		if(++nextTask == SMARTS->getTotalTasks())
 			nextTask = 1;
